organize chat presentation files
- Move BasePagingSource to 'shared' package
- Rename 'MainScreen.kt' to 'ChatsScreen.kt' and move to 'chats' package
- Remove unused 'ContactsScreen.kt' (the one in 'screens' package)
- Update package names for files moved from 'screens' to 'screen' sub-packages
- Update navigation routes and NavHost configuration
