+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/navigation/ChatNavHost.kt
@@ -7,8 +7,9 @@ import androidx.compose.ui.Modifier
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import net.thechance.mena.core_chat.presentation.screens.ContactsScreen
-import net.thechance.mena.core_chat.presentation.screens.MainScreen
+import net.thechance.mena.core_chat.presentation.screen.contacts.ContactsScreen
+import net.thechance.mena.core_chat.presentation.screen.syncContacts.SyncContactsScreen
+import net.thechance.mena.core_chat.presentation.screen.chats.ChatsScreen
 
 @Composable
 fun ChatNavHost() {
@@ -21,15 +22,11 @@ fun ChatNavHost() {
         NavHost(
             modifier = Modifier.fillMaxSize(),
             navController = navController,
-            startDestination = MainRoute,
+            startDestination = ChatsRoute,
         ) {
-            composable<MainRoute> {
-                MainScreen()
-            }
-
-            composable<ContactsRoute> {
-                ContactsScreen()
-            }
+            composable<ChatsRoute> { ChatsScreen() }
+            composable<ContactsRoute> { ContactsScreen() }
+            composable<SyncContactsRoute> { SyncContactsScreen() }
         }
     }
 }
